@@ -2,7 +2,7 @@ import './index.css'
 
 import { useToggle } from 'src/hooks'
 import { useTranslation } from 'react-i18next'
-import Info from '@material-ui/icons/Info'
+import Info from '@mui/icons-material/Info'
 import React from 'react'
 
 interface Props {
@@ -31,8 +31,8 @@ export default function InfoBox({ children, text }: Props) {
         }}
         data-testid="infoboxSpan"
       >
-        <p>{t(text)}</p>
         <Info className="material-icons" />
+        <p>{t(text)}</p>
       </a>
       <div
         style={{ display: isHidden ? 'none' : 'block' }}
