@@ -4,17 +4,19 @@ import { Path, WineInstallation } from 'src/types'
 import { useTranslation } from 'react-i18next'
 import { InfoBox, ToggleSwitch, SvgButton } from 'src/components/UI'
 
-import AddBoxIcon from '@material-ui/icons/AddBox'
+import AddBoxIcon from '@mui/icons-material/AddBox'
 import ContextProvider from 'src/state/ContextProvider'
-import CreateNewFolder from '@material-ui/icons/CreateNewFolder'
-import RemoveCircleIcon from '@material-ui/icons/RemoveCircle'
+import CreateNewFolder from '@mui/icons-material/CreateNewFolder'
+import RemoveCircleIcon from '@mui/icons-material/RemoveCircle'
 import classNames from 'classnames'
+import { Tooltip } from '@mui/material'
 
 const { ipcRenderer } = window.require('electron')
 
 interface Props {
   altWine: WineInstallation[]
   autoInstallDxvk: boolean
+  autoInstallVkd3d: boolean
   customWinePaths: string[]
   isDefault: boolean
   maxSharpness: number
@@ -27,6 +29,7 @@ interface Props {
   setFsrSharpness: (value: number) => void
   setWineVersion: (wine: WineInstallation) => void
   toggleAutoInstallDxvk: () => void
+  toggleAutoInstallVkd3d: () => void
   toggleFSR: () => void
   toggleResizableBar: () => void
   wineCrossoverBottle: string
@@ -48,9 +51,11 @@ export default function WineSettings({
   wineVersion,
   altWine,
   toggleAutoInstallDxvk,
+  toggleAutoInstallVkd3d,
   enableFSR,
   toggleFSR,
   autoInstallDxvk,
+  autoInstallVkd3d,
   customWinePaths,
   setCustomWinePaths,
   wineCrossoverBottle,
@@ -140,7 +145,13 @@ export default function WineSettings({
                   )
               }
             >
-              <CreateNewFolder data-testid="addWinePrefix" />
+              <CreateNewFolder
+                data-testid="addWinePrefix"
+                titleAccess={t(
+                  'toolbox.settings.wineprefix',
+                  'Select a Folder for new Wine Prefixes'
+                )}
+              />
             </SvgButton>
           </span>
         </span>
@@ -172,7 +183,13 @@ export default function WineSettings({
                   )
               }
             >
-              <CreateNewFolder data-testid="addWinePrefix" />
+              <CreateNewFolder
+                data-testid="addWinePrefix"
+                titleAccess={t(
+                  'toolbox.settings.default-wineprefix',
+                  'Select the default prefix folder for new configs'
+                )}
+              />
             </SvgButton>
           </span>
         </span>
@@ -197,28 +214,38 @@ export default function WineSettings({
             </select>
             <div className="iconsWrapper">
               <SvgButton onClick={() => removeCustomPath()}>
-                <RemoveCircleIcon
-                  data-testid="removeWinePath"
-                  style={{
-                    color: selectedPath
-                      ? 'var(--danger)'
-                      : 'var(--background-darker)',
-                    cursor: selectedPath ? 'pointer' : ''
-                  }}
-                  fontSize="large"
-                  titleAccess={t('tooltip.removepath', 'Remove Path')}
-                />
+                <Tooltip
+                  title={t('tooltip.removepath', 'Remove Path') as string}
+                  placement="bottom"
+                  arrow
+                >
+                  <RemoveCircleIcon
+                    data-testid="removeWinePath"
+                    style={{
+                      color: selectedPath
+                        ? 'var(--danger)'
+                        : 'var(--background-darker)',
+                      cursor: selectedPath ? 'pointer' : ''
+                    }}
+                    fontSize="large"
+                  />
+                </Tooltip>
               </SvgButton>{' '}
               <SvgButton
                 onClick={() => selectCustomPath()}
                 className={`is-primary`}
               >
-                <AddBoxIcon
-                  data-testid="addWinePath"
-                  style={{ color: 'var(--success)', cursor: 'pointer' }}
-                  fontSize="large"
-                  titleAccess={t('tooltip.addpath', 'Add New Path')}
-                />
+                <Tooltip
+                  title={t('tooltip.addpath', 'Add New Path') as string}
+                  placement="bottom"
+                  arrow
+                >
+                  <AddBoxIcon
+                    data-testid="addWinePath"
+                    style={{ color: 'var(--success)', cursor: 'pointer' }}
+                    fontSize="large"
+                  />
+                </Tooltip>
               </SvgButton>
             </div>
           </span>
@@ -243,6 +270,23 @@ export default function WineSettings({
           ))}
         </select>
       </span>
+      <InfoBox text="infobox.help">
+        <span>{t('help.wine.part1')}</span>
+        <ul>
+          <i>
+            <li>~/.config/heroic/tools/wine</li>
+            <li>~/.config/heroic/tools/proton</li>
+            <li>~/.steam/root/compatibilitytools.d</li>
+            <li>~/.steam/steamapps/common</li>
+            <li>~/.local/share/lutris/runners/wine</li>
+            <li>~/.var/app/com.valvesoftware.Steam (Steam Flatpak)</li>
+            <li>/usr/share/steam</li>
+            <li>Everywhere on the system (CrossOver Mac)</li>
+            <li>/opt/cxoffice (CrossOver Linux)</li>
+          </i>
+        </ul>
+        <span>{t('help.wine.part2')}</span>
+      </InfoBox>
       {wineVersion.name.includes('CrossOver') && (
         <span className="setting">
           <span className={classNames('settingText', { isRTL: isRTL })}>
@@ -262,7 +306,6 @@ export default function WineSettings({
       {isLinux && !isProton && (
         <span className="setting">
           <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-            {t('setting.autodxvk', 'Auto Install/Update DXVK on Prefix')}
             <ToggleSwitch
               value={autoInstallDxvk}
               handleChange={() => {
@@ -278,15 +321,38 @@ export default function WineSettings({
                 'Auto Install/Update DXVK on Prefix'
               )}
             />
+            <span>
+              {t('setting.autodxvk', 'Auto Install/Update DXVK on Prefix')}
+            </span>
+          </span>
+        </span>
+      )}
+      {isLinux && !isProton && (
+        <span className="setting">
+          <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
+            <ToggleSwitch
+              value={autoInstallVkd3d}
+              handleChange={() => {
+                const action = autoInstallVkd3d ? 'restore' : 'backup'
+                ipcRenderer.send('toggleVKD3D', [
+                  { winePrefix, winePath: wineVersion.bin },
+                  action
+                ])
+                return toggleAutoInstallVkd3d()
+              }}
+              title={t(
+                'setting.autovkd3d',
+                'Auto Install/Update VKD3D on Prefix'
+              )}
+            />
+            <span>
+              {t('setting.autovkd3d', 'Auto Install/Update VKD3D on Prefix')}
+            </span>
           </span>
         </span>
       )}
       <span className="setting">
         <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-          {t(
-            'setting.enableFSRHack',
-            'Enable FSR Hack (Wine version needs to support it)'
-          )}
           <ToggleSwitch
             value={enableFSR || false}
             handleChange={toggleFSR}
@@ -295,12 +361,17 @@ export default function WineSettings({
               'Enable FSR Hack (Wine version needs to support it)'
             )}
           />
+          <span>
+            {t(
+              'setting.enableFSRHack',
+              'Enable FSR Hack (Wine version needs to support it)'
+            )}
+          </span>
         </span>
       </span>
       {enableFSR && (
         <span className="setting">
           <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-            {t('setting.FsrSharpnessStrenght', 'FSR Sharpness Strength')}
             <select
               data-testid="setMaxRecentGames"
               onChange={(event) => setFsrSharpness(Number(event.target.value))}
@@ -311,6 +382,9 @@ export default function WineSettings({
                 <option key={n + 1}>{n + 1}</option>
               ))}
             </select>
+            <span>
+              {t('setting.FsrSharpnessStrenght', 'FSR Sharpness Strength')}
+            </span>
           </span>
         </span>
       )}
@@ -318,10 +392,6 @@ export default function WineSettings({
         <>
           <span className="setting">
             <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-              {t(
-                'setting.resizableBar',
-                'Enable Resizable BAR (NVIDIA RTX only)'
-              )}
               <ToggleSwitch
                 value={enableResizableBar || false}
                 handleChange={toggleResizableBar}
@@ -330,49 +400,38 @@ export default function WineSettings({
                   'Enable Resizable BAR (NVIDIA RTX only)'
                 )}
               />
+              <span>
+                {t(
+                  'setting.resizableBar',
+                  'Enable Resizable BAR (NVIDIA RTX only)'
+                )}
+              </span>
             </span>
           </span>
           <span className="setting">
             <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-              {t('setting.esync', 'Enable Esync')}
               <ToggleSwitch
                 value={enableEsync || false}
                 handleChange={toggleEsync}
                 dataTestId="esyncToggle"
                 title={t('setting.esync', 'Enable Esync')}
               />
+              <span>{t('setting.esync', 'Enable Esync')}</span>
             </span>
           </span>
           <span className="setting">
             <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-              {t('setting.fsync', 'Enable Fsync')}
               <ToggleSwitch
                 value={enableFsync || false}
                 handleChange={toggleFsync}
                 dataTestId="fsyncToggle"
                 title={t('setting.fsync', 'Enable Fsync')}
               />
+              <span>{t('setting.fsync', 'Enable Fsync')}</span>
             </span>
           </span>
         </>
       )}
-      <InfoBox text="infobox.help">
-        <span>{t('help.wine.part1')}</span>
-        <ul>
-          <i>
-            <li>~/.config/heroic/tools/wine</li>
-            <li>~/.config/heroic/tools/proton</li>
-            <li>~/.steam/root/compatibilitytools.d</li>
-            <li>~/.steam/steamapps/common</li>
-            <li>~/.local/share/lutris/runners/wine</li>
-            <li>~/.var/app/com.valvesoftware.Steam (Steam Flatpak)</li>
-            <li>/usr/share/steam</li>
-            <li>Everywhere on the system (CrossOver Mac)</li>
-            <li>/opt/cxoffice (CrossOver Linux)</li>
-          </i>
-        </ul>
-        <span>{t('help.wine.part2')}</span>
-      </InfoBox>
     </>
   )
 }
